+++ Example/ObjCDemo/ObjCDemo/DetailViewController.h
@@ -7,9 +7,17 @@
 //
 
 #import <UIKit/UIKit.h>
+#import "XLForm.h"
 /* Google Native SignIn
 #import <Google/SignIn.h>
 */
-@interface DetailViewController : UIViewController
+/* Twitter Native SignIn
+#import <TwitterKit/TwitterKit.h>
+*/
+
+@interface DetailViewController : XLFormViewController
+
+
++ (NSArray<NSString *>*) countries;
 
 @end